requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.6
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.8.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
kombu==5.2.4
MarkupSafe==2.1.1
platformdirs==2.5.4
prompt-toolkit==3.0.36
python-dateutil==2.8.2
pytz==2022.6
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tzdata==2022.6
vine==5.0.0
virtualenv==20.16.7
wcwidth==0.2.6
Werkzeug==2.2.2
